レイアウト

+++ app/controllers/public/comments_controller.rb
@@ -1,4 +1,6 @@
 class Public::CommentsController < ApplicationController
+  before_action :authenticate_user!, only: [:create]
+  
   def new
   end
 
@@ -12,7 +14,14 @@ class Public::CommentsController < ApplicationController
 
   def destroy
     comment = Comment.find(params[:id])
-    comment.destroy
+    if comment.user == current_user
+      comment.destroy
+      flash[:notice] = 'コメントが削除されました'
+    else
+      flash[:alert] = '権限がありません'
+    end
+    redirect_to request.referer
+    # その場に留まる
   end
 
   private
